MSPSDS-538: Move patching to an initializer, fix problem with hot reloading
