+++ web/app/models/investigation.rb
@@ -51,7 +51,6 @@ class Investigation < ApplicationRecord
   after_create :create_audit_activity_for_case
 
   def as_indexed_json(*)
-    expose_file_data_in_attachment
     as_json(
       methods: :pretty_id,
       only: %i[question_title description],
@@ -202,24 +201,6 @@ private
     first_product = products.first
     if products.all? { |product| product[property_name] == first_product[property_name] }
       first_product[property_name]
-    end
-  end
-
-  def expose_file_data_in_attachment
-    return if ActiveStorage::Attachment.method_defined?(:title)
-
-    ActiveStorage::Attachment.class_eval do
-      def title
-        blob.metadata["title"]
-      end
-
-      def filename
-        blob.filename
-      end
-
-      def description
-        blob.metadata["description"]
-      end
     end
   end
 end
